ajout test statistiques dans joueurServiceTets et corrections pour qu'il s'execute
